Several tiny improvements.
Added a tilt indicator at the top-right. Should give the user a hint
that it works by tilting the device. It is only shown if the required
HTML5/JavaScript APIs work.

Added a link at the bottom-right to the README.

Added Google Analytics. I should have added it BEFORE Simone retweeted
it.